fix(quality): land ReSharper remediation and pending main changes (#623)
## Changes

Land the reviewed ReSharper remediation together with the four existing unpublished main commits for combat presentation, Slice C command review/coordination and acceptance, and Slice G planning.

The remediation fixes 1,363 findings, records 956 reviewed exceptions and seven owned integration follow-ups, and adds a pinned all-framework ReSharper CLI regression gate. It corrects deep rollback restoration and exception parameter names while retaining public DTO contracts and runtime guards. The remediation leaves DelegationBridge, CatalogWriteGate write paths, and Baltic v2 goldens unchanged.

## Verification

- Debug and Release builds: zero warnings/errors; 3,226 tests passed in each configuration.
- Headless PlayMode smoke: 24 passed; locked ReplayGoldenSuite: 6 passed; replay filter: 17 passed.
- Tooling tests: 19 passed; catalog checks: 67 passed; proprietary database gate passed.
- ReSharper CLI 2026.2.1, all frameworks with SWEA: zero new, open, reintroduced, or ambiguous findings.
- Release netstandard2.1 Unity plugin refresh: all 14 DLL hashes verified.
- Baltic v2 hash 17144800277401907079 preserved.

## Unity/C# architecture verdict (remediation scope)

**Verdict: PASS**

All applicable architecture gates, headless tests, last-mile smoke tests, and adapter plugin refresh checks passed. No architecture defect or waiver is open. The final repository-wide analyzer wrapper also passed with zero new, reintroduced, ambiguous, or open findings.

Presentation boundary: ADR-010 sections 2-3, ADR-007, ADR-001. Full review: docs/superpowers/reviews/2026-09-10-resharper-architecture-review.md. Closeout and evidence manifest: docs/superpowers/reviews/2026-09-10-resharper-closeout.md and 2026-09-10-resharper-verification.json.

## CI portability fixes

Exclude the existing invalid roadmap symlink from the Windows inspection checkout. Run test projects sequentially in the shared blocking gate so wall-clock UI budgets do not compete with other assemblies; all tests and 100 ms assertions remain enabled. The sequential local run passed all 3,226 tests.

+++ src/ProjectAegis.Delegation.Tests/IdentityClass/IdentityClassProjectionTests.cs
@@ -117,7 +117,7 @@ public sealed class IdentityClassProjectionTests
         var log = new DecisionLog();
         log.AppendContactChange(Change(1, "c1", "hostile-1", "Unknown", "Detected"));
 
-        var staleTick = 1UL + (ulong)ContactProvenanceProjection.DefaultStaleThresholdTicks + 1;
+        var staleTick = 1UL + ContactProvenanceProjection.DefaultStaleThresholdTicks + 1;
         var snapshot = IdentityClassProjection.Project(log, currentSimTick: staleTick);
 
         var row = snapshot.Rows[0];
